updated storing system, changed to java 8, preparations for v2
