+++ umbrellapanel-chief/src/main/java/ms/idrea/umbrellapanel/chief/gameserver/UmbrellaServerManager.java
@@ -6,12 +6,15 @@ import java.io.IOException;
 import java.io.Reader;
 import java.io.Writer;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
+
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 
 import lombok.ToString;
-
 import ms.idrea.umbrellapanel.api.chief.PanelUserDatabase;
 import ms.idrea.umbrellapanel.api.chief.WorkerManager;
 import ms.idrea.umbrellapanel.api.gameserver.ManagedServer;
@@ -27,9 +30,10 @@ import ms.idrea.umbrellapanel.net.messages.UpdateMultiInstanceServerMessage;
 @ToString(of = { "servers" })
 public class UmbrellaServerManager implements ServerManager {
 
-	private PanelUserDatabase panelUserDatabase;
-	private WorkerManager workerManager;
-	private ConcurrentMap<Integer, ManagedServer> servers = new ConcurrentHashMap<>();
+	private static final JsonParser PARSER = new JsonParser();
+	private final PanelUserDatabase panelUserDatabase;
+	private final WorkerManager workerManager;
+	private final List<ManagedServer> servers = Collections.synchronizedList(new ArrayList<>());
 	private int nextId = 0;
 
 	public UmbrellaServerManager(PanelUserDatabase panelUserDatabase, WorkerManager workerManager) {
@@ -44,7 +48,7 @@ public class UmbrellaServerManager implements ServerManager {
 
 	@Override
 	public List<ManagedServer> getAllServers() {
-		return new ArrayList<>(servers.values());
+		return Collections.unmodifiableList(servers);
 	}
 
 	@Override
@@ -54,25 +58,30 @@ public class UmbrellaServerManager implements ServerManager {
 
 	@Override
 	// only used for NEW servers.
-	public UmbrellaSingleInstanceGameServer createSingleInstanceServer(Address address, String startCommand, int workerId) {
-		UmbrellaSingleInstanceGameServer server = new UmbrellaSingleInstanceGameServer(getNextId(), workerId, address, startCommand, workerManager, panelUserDatabase);
+	public UmbrellaSingleInstanceGameServer createSingleInstanceServer(Address address, String startCommand,
+			int workerId) {
+		UmbrellaSingleInstanceGameServer server = new UmbrellaSingleInstanceGameServer(getNextId(), workerId, address,
+				startCommand, workerManager, panelUserDatabase);
 		if (server.getOnlineWorker() == null) {
 			throw new IllegalStateException("Worker is offline!");
 		}
-		servers.put(server.getId(), server);
-		server.getWorkerOrThrow().send(new UpdateGameServerMessage(UpdateGameServerMessage.Action.CREATE, server.getId(), server.getAddress(), server.getStartCommand()));
+		servers.add(server);
+		server.getWorkerOrThrow().send(new UpdateGameServerMessage(UpdateGameServerMessage.Action.CREATE,
+				server.getId(), server.getAddress(), server.getStartCommand()));
 		return server;
 	}
 
 	@Override
 	// only used for NEW servers.
 	public UmbrellaMultiInstanceGameServer createMultiInstanceServer(String startCommand, int workerId) {
-		UmbrellaMultiInstanceGameServer server = new UmbrellaMultiInstanceGameServer(getNextId(), workerId, startCommand, workerManager, panelUserDatabase, 0);
+		UmbrellaMultiInstanceGameServer server = new UmbrellaMultiInstanceGameServer(getNextId(), workerId,
+				startCommand, workerManager, panelUserDatabase, 0);
 		if (server.getOnlineWorker() == null) {
 			throw new IllegalStateException("Worker is offline!");
 		}
-		servers.put(server.getId(), server);
-		server.getWorkerOrThrow().send(new UpdateMultiInstanceServerMessage(UpdateGameServerMessage.Action.CREATE, server.getId(), server.getStartCommand()));
+		servers.add(server);
+		server.getWorkerOrThrow().send(new UpdateMultiInstanceServerMessage(UpdateGameServerMessage.Action.CREATE,
+				server.getId(), server.getStartCommand()));
 		return server;
 	}
 
@@ -85,85 +94,78 @@ public class UmbrellaServerManager implements ServerManager {
 	@Override
 	public void save(Writer out) throws IOException {
 		BufferedWriter writer = new BufferedWriter(out);
-		writer.write(String.valueOf(nextId));
-		writer.newLine();
-		writer.write(String.valueOf(servers.size()));
-		writer.newLine();
-		for (int id : servers.keySet()) {
-			GameServer s = getServer(id);
-			writer.write(String.valueOf(s.getId()));
-			writer.newLine();
-			writer.write(String.valueOf(((ManagedServer) s).getWorker().getId()));
-			writer.newLine();
-			writer.write(String.valueOf(s.getStartCommand()));
-			writer.newLine();
-			writer.write(String.valueOf(((ManagedServer) s).getName()));
-			writer.newLine();
-			if (s instanceof SingleInstanceServer) {
-				writer.write(String.valueOf(0));
-				writer.newLine();
-				SingleInstanceServer server = (SingleInstanceServer) s;
-				writer.write(String.valueOf(server.getAddress().getHost()));
-				writer.newLine();
-				writer.write(String.valueOf(server.getAddress().getPort()));
-				writer.newLine();
-			} else if (s instanceof MultiInstanceServer) {
-				writer.write(String.valueOf(1));
-				writer.newLine();
-				MultiInstanceServer server = (MultiInstanceServer) s;
-				writer.write(String.valueOf(server.getNextInstanceId()));
-				writer.newLine();
-				writer.write(String.valueOf(server.getInstances().size()));
-				writer.newLine();
-				for (ServerInstance instance : server.getInstances()) {
-					writer.write(String.valueOf(instance.getId()));
-					writer.newLine();
-					writer.write(String.valueOf(instance.getAddress().getHost()));
-					writer.newLine();
-					writer.write(String.valueOf(instance.getAddress().getPort()));
-					writer.newLine();
+		JsonObject serverSaveData = new JsonObject();
+		serverSaveData.addProperty("nextId", nextId);
+		JsonArray serversArray = new JsonArray();
+		for (GameServer server : servers) {
+			JsonObject obj = new JsonObject();
+			obj.addProperty("id", server.getId());
+			obj.addProperty("worker", String.valueOf(((ManagedServer) server).getWorker().getId()));
+			obj.addProperty("startCommand", server.getStartCommand());
+			if (server instanceof SingleInstanceServer) {
+				obj.addProperty("type", "SingleInstanceServer");
+				SingleInstanceServer s = (SingleInstanceServer) server;
+				obj.addProperty("address", s.getAddress().getHost());
+				obj.addProperty("port", s.getAddress().getPort());
+			} else {
+				obj.addProperty("type", "MultiInstanceServer");
+				MultiInstanceServer s = (MultiInstanceServer) server;
+				obj.addProperty("nextInstanceId", s.getNextInstanceId());
+				JsonArray instancesArray = new JsonArray();
+				for (ServerInstance instance : s.getInstances()) {
+					JsonObject instanceObject = new JsonObject();
+					instanceObject.addProperty("id", instance.getId());
+					instanceObject.addProperty("address", instance.getAddress().getHost());
+					instanceObject.addProperty("port", instance.getAddress().getPort());
+					instancesArray.add(instanceObject);
 				}
+				obj.add("instances", instancesArray);
 			}
+			serversArray.add(obj);
 		}
+		serverSaveData.add("servers", serversArray);
+		writer.write(serverSaveData.toString());
 		writer.flush();
 	}
 
 	@Override
 	public void load(Reader in) throws IOException {
 		BufferedReader reader = new BufferedReader(in);
-		nextId = Integer.valueOf(reader.readLine());
-		int serverSize = Integer.valueOf(reader.readLine());
-		servers = new ConcurrentHashMap<>(serverSize);
-		for (int i = 0; i < serverSize; i++) {
-			ManagedServer server;
-			int id = Integer.valueOf(reader.readLine());
-			int workerId = Integer.valueOf(reader.readLine());
-			String startCommand = reader.readLine();
-			String name = reader.readLine();
-			int type = Integer.valueOf(reader.readLine());
-			if (type == 0) {
-				// single instance
-				String host = reader.readLine();
-				int port = Integer.valueOf(reader.readLine());
-				server = new UmbrellaSingleInstanceGameServer(id, workerId, new Address(host, port), startCommand, workerManager, panelUserDatabase);
-			} else if (type == 1) {
-				// multi instance
-				//
-				int nextInstanceId = Integer.valueOf(reader.readLine());
-				int instanceCount = Integer.valueOf(reader.readLine());
-				server = new UmbrellaMultiInstanceGameServer(id, workerId, startCommand, workerManager, panelUserDatabase, nextInstanceId);
-				for (int j = 0; j < instanceCount; j++) {
-					int instanceId = Integer.valueOf(reader.readLine());
-					String host = reader.readLine();
-					int port = Integer.valueOf(reader.readLine());
-					((MultiInstanceServer) server).addInstance(instanceId, new Address(host, port));
+		String readed = reader.readLine();
+		if (readed == null) {
+			return;
+		}
+		JsonObject obj = PARSER.parse(readed).getAsJsonObject();
+		this.nextId = obj.get("nextId").getAsInt();
+		JsonArray serversArray = obj.get("servers").getAsJsonArray();
+		for (JsonElement element : serversArray) {
+			ManagedServer server = null;
+			JsonObject serverData = (JsonObject) element;
+			int id = serverData.get("id").getAsInt();
+			int workerId = serverData.get("worker").getAsInt();
+			String startCommand = serverData.get("startCommand").getAsString();
+			if (serverData.get("type").getAsString().equals("SingleInstanceServer")) {
+				Address address = new Address(serverData.get("address").getAsString(),
+						serverData.get("port").getAsInt());
+				server = new UmbrellaSingleInstanceGameServer(id, workerId, address, startCommand, workerManager,
+						panelUserDatabase);
+			} else if (serverData.get("type").getAsString().equals("MultiInstanceServer")) {
+				int nextInstanceId = serverData.get("nextInstanceId").getAsInt();
+				server = new UmbrellaMultiInstanceGameServer(id, workerId, startCommand, workerManager,
+						panelUserDatabase, nextInstanceId);
+				JsonArray instancesArray = serverData.get("instances").getAsJsonArray();
+				for (JsonElement instanceElement : instancesArray) {
+					JsonObject instanceObject = (JsonObject) instanceElement;
+					int instanceId = instanceObject.get("id").getAsInt();
+					Address address = new Address(instanceObject.get("address").getAsString(),
+							instanceObject.get("port").getAsInt());
+					((MultiInstanceServer) server).addInstance(instanceId, address);
 				}
-			} else {
-				System.out.println("Unsupported type: " + type);
-				continue;
 			}
-			server.setName(name);
-			servers.put(id, server);
+			if (server == null) {
+				throw new RuntimeException("Error while loading server: " + element);
+			}
+			servers.add(server);
 		}
 	}
 }
